refactor: Change README and update LICENSE to include myself
This project is a fork of the Semaphore UI project and its purpose is
to allow users to automate tasks on a daily basis.

I will use this project as a simple infrastructure for my Linux IoT
devices but technically this can be used where any other timetable based
automation is needed.